Restore yoast/components.
