+++ packages/components/src/data-model/data-model.css
@@ -1,7 +1,6 @@
 .yoast-data-model {
     padding: 0;
     list-style: none;
-    width: 275px;
 }
 
 .yoast-data-model li {
